refactor(settings): merge declarative bindings
