+++ apps/desktop/tests/stores/settingsDefinitions.test.ts
@@ -1,18 +1,27 @@
-import type { GeneralSettingKey } from '@services/EventService';
-import { describe, expect, it } from 'vitest';
+import type { GeneralSettingKey, SettingsGeneralUpdatedEvent } from '@services/EventService';
+import { describe, expect, it, vi } from 'vitest';
+import { ref } from 'vue';
 
+import { DEFAULT_BROWSER_SETTINGS } from '@/config/browserSettings';
+import { DEFAULT_SEARCH_SETTINGS } from '@/config/searchSettings';
 import { JSON_SETTINGS_SECTIONS } from '@/config/settingsRegistry';
 import {
     applyGeneralSettingValue,
     applyPersistedGeneralSettingValue,
     cloneGeneralSettingsSnapshot,
     createDefaultGeneralSettings,
+    createGeneralSettingsComputedRefs,
+    createGeneralSettingUpdaters,
+    GENERAL_SETTING_COMPUTED_BINDINGS,
     GENERAL_SETTING_DEFINITIONS,
+    GENERAL_SETTING_UPDATER_BINDINGS,
     getGeneralSettingDefinition,
     getGeneralSettingEventValue,
     JSON_GENERAL_SETTING_DEFINITIONS,
     serializeGeneralSetting,
 } from '@/stores/settingsDefinitions';
+
+type GeneralSettingValue = SettingsGeneralUpdatedEvent['value'];
 
 const EXPECTED_GENERAL_SETTING_KEYS: GeneralSettingKey[] = [
     'global_shortcut',
@@ -102,5 +111,74 @@ describe('settings definitions', () => {
 
         expect(settings.browserSettings.permissions.navigate).not.toBe('deny');
         expect(settings.searchSettings.providers.anysearch.enabled).toBe(true);
+    });
+
+    it('declares stable computed and updater bindings in the settings registry', () => {
+        expect(GENERAL_SETTING_COMPUTED_BINDINGS.map((binding) => binding.exposedName)).toEqual([
+            'outputScrollBehavior',
+            'globalShortcut',
+            'searchWindowSizePreset',
+            'searchWindowDefaultSize',
+            'language',
+            'appUpdateChannel',
+            'appUpdateAutoCheck',
+            'appUpdateLastCheckedAt',
+            'browserSettings',
+            'searchSettings',
+        ]);
+
+        expect(GENERAL_SETTING_UPDATER_BINDINGS.map((binding) => binding.exposedName)).toEqual([
+            'updateGlobalShortcut',
+            'updateStartOnBoot',
+            'updateStartMinimized',
+            'updateOutputScrollBehavior',
+            'updateSearchWindowSizePreset',
+            'updateLanguage',
+            'updateAppUpdateChannel',
+            'updateAppUpdateAutoCheck',
+            'updateAppUpdateLastCheckedAt',
+            'updateBrowserSettings',
+            'updateSearchSettings',
+        ]);
+    });
+
+    it('creates computed refs from declared state bindings', () => {
+        const settings = ref(createDefaultGeneralSettings());
+        const computedRefs = createGeneralSettingsComputedRefs(settings);
+
+        expect(computedRefs.globalShortcut.value).toBe('Alt+Space');
+
+        settings.value.globalShortcut = 'Ctrl+Space';
+
+        expect(computedRefs.globalShortcut.value).toBe('Ctrl+Space');
+    });
+
+    it('creates updater methods from declared key bindings', async () => {
+        const updateSetting = vi.fn(async (key: GeneralSettingKey, value: GeneralSettingValue) => {
+            void key;
+            void value;
+        });
+        const updaters = createGeneralSettingUpdaters(updateSetting);
+
+        await updaters.updateGlobalShortcut('Ctrl+Space');
+        await updaters.updateLanguage('en-US');
+        await updaters.updateBrowserSettings({
+            ...DEFAULT_BROWSER_SETTINGS,
+            defaultHomepage: 'https://example.test',
+        });
+        await updaters.updateSearchSettings(DEFAULT_SEARCH_SETTINGS);
+
+        expect(updateSetting).toHaveBeenNthCalledWith(1, 'global_shortcut', 'Ctrl+Space');
+        expect(updateSetting).toHaveBeenNthCalledWith(2, 'language', 'en-US');
+        expect(updateSetting).toHaveBeenNthCalledWith(
+            3,
+            'browser_settings',
+            expect.stringContaining('https://example.test')
+        );
+        expect(updateSetting).toHaveBeenNthCalledWith(
+            4,
+            'search_settings',
+            expect.stringContaining('providers')
+        );
     });
 });
